+++ packages/interchain/interchain-daemon/src/interchain_env.rs
@@ -28,6 +28,7 @@ use std::time::Duration;
 use tokio::runtime::Handle;
 
 /// Represents a set of locally running blockchain nodes and a Hermes relayer.
+#[derive(Clone)]
 pub struct DaemonInterchainEnv<C: ChannelCreator = ChannelCreationValidator> {
     /// Daemons indexable by network id, i.e. "juno-1", "osmosis-2", ...
     daemons: HashMap<NetworkId, Daemon>,
@@ -89,10 +90,17 @@ impl<C: ChannelCreator> DaemonInterchainEnv<C> {
         mnemonic: Option<impl ToString>,
     ) -> IcDaemonResult<()> {
         let mut daemon_builder = Daemon::builder();
-        let daemon_builder = daemon_builder.chain(chain_data.clone()).handle(runtime);
+        let mut daemon_builder = daemon_builder.chain(chain_data.clone()).handle(runtime);
 
-        let daemon_builder = if let Some(mn) = mnemonic {
+        daemon_builder = if let Some(mn) = mnemonic {
             daemon_builder.mnemonic(mn)
+        } else {
+            daemon_builder
+        };
+
+        // State is shared between daemons, so if a daemon already exists, we use its state
+        daemon_builder = if let Some(daemon) = self.daemons.values().next() {
+            daemon_builder.state(daemon.state())
         } else {
             daemon_builder
         };